Centralize resource fetching to the App component including fixing JWT error via Auth
