+++ src/layout/mainContent/characters/Characters.js
@@ -1,12 +1,7 @@
 import { useDispatch, useSelector } from "react-redux";
-import { useEffect } from "react";
-
 import { Table } from "../../reusables/Table";
-import { initDatasetThunk } from "../../../redux/dataSlice";
 
 export const Characters = () => {
-  const dispatch = useDispatch();
-  const userData = useSelector((state) => state.login.authInfo.user);
   const allData = useSelector((state) => state.data.characters);
 
   const tableCols = [
@@ -27,18 +22,6 @@ export const Characters = () => {
     columns: tableCols,
     type: "characters",
   };
-
-  // Run this request once when the component first renders to request characters
-  useEffect(async () => {
-    const promise = await dispatch(
-      initDatasetThunk({
-        jwt: userData.signInUserSession.idToken.jwtToken,
-      })
-    );
-    return () => {
-      promise.abort();
-    };
-  }, [initDatasetThunk]);
 
   return (
     <>
